relocate clientRequest and asset download
